dramatically simplify CSS/rendering for timeslice/timeslicegroup
Uses flex box for rendering of arbitrary number of time slice groups.
Removes grid lines from the gutter on the left.  This solves 2 problems:
when there are more than 2 grid lines, they force the top slice with
the label to be larger than all the others.  In addition, it is closer
to what google calendar looks like, so people will find it more familiar