Add an indicator that represents the current time
In the week and day view a horizontal line represents the current time.
Apple Calendar and other apps use this concept to provide the current
time as a subtle hint.

The Y position of line is calculated in the TimeGrid by using a date
diff between this.props.max and this.props.min, respectively the
diff between current date and this.props.min.
To convert the result to a pixel value, the height of the time gutter
is used as a reference.
The position of the time indicator is calulated and updated every
minute using setInterval().

Please let me know if you want this feature to be toggled using a
prop.