Pass 'selected' prop through to All Day Events
The 'selected' prop was not being passed through to the header Event Rows for All Day events in Month/Week views, which meant that any logic which relied on that prop (special styling for selected events, etc.) wouldn't work. Fix is easy: just make sure to pass that prop down.