Add a class with the weekday
This change adds a class with the weekday, similar to the
`rbc-today`. The purpose of this class is to allow us to customize the
cells by day of the week. For instance, to customize weekends we could
do:

```
.rbc-weekday-0,
.rbc-weekday-6 {
  background: red;
}
```

I did not write any default styles to avoid breaking the API or the
default behavior and appearance. Hopefully, this will make it easier to
merge this change.