ResizableEventLR Component
Summary: I created <ResizableEventLR /> to handle resizing events in the month view. If <EventCell resizable /> then it will render the resizable component.

I added styles so that I could see what is going on. There is a red bg. Sorry, not sorry.

There is no functionality at the moment but I started building it out. There is certainly more to be worked on here.