Fix Deselect Bug
Summary: This commit fixes a bug in how RBC selects events. Once an event is selected it is held in state. With no deselect or blur event, when the user clicks away from the event it is still in a selected state. These changes allow the event to be focused and when a blur event occurs it allows the event to not be selected.