Context Menu: Make More Dynamic
Summary: In order to make the context menu more dynamic I am using connectMenu from react-contextmenu. Then I pulled out the entire context menu from react-big-calendar so that it can be dropped in from itv-calendar. I want to make react-big-calendar as dumb as possible so that changes that need to be made are made within itv-calendar and not react-big-calendar.