UI: Elevate rbc-day-bg-wrapper when isDragging
Summary: This small change to the styles allows a user to drop an event anywhere on a day cell and not have to "find" a place to drop an event in a day cell. Previously, if there were many events on the same day, the user would have to drop the cursor on the top part of the day cell to resize the event. The style update elevates the z-index of the rbc-day-bg-wrapper to be at 10 when the calendar is in a drag state so that the day cell is above the event, not behind.