Support point-in-time events in the Agenda view
Some events may have the same date/time, and are meant to represent
point-in-time events, such as deadlines. These are not all day
events, but they also should not be represnted as
`${startTime} - ${endTime}` since the values will be the same.

This will cause an event that is not marked as all day, and where
the start and end are equal, to show as:

```js
// old
`${startTime} - ${endTime}`

// new
`${startTime}`
```