Allow for custom titles in the toolbar
Currently, even when building a custom toolbar, the `label` prop
passed to it is not controllable. Further, for someone who wants
to modify just the `label`, building a custom toolbar component is
overly complex.

This allows for the user to pass a new prop, `getLabel`, which
allows for customizing the label passed to the toolbar. The prop
is a function which is given access to the current date, the
current view, and the view's length (when needed).

This also adds the `Custom Header` example to the examples list.
This page existed but was not listed. This fixes that.