Fix display of dates around daylight savings (#7)
A display bug showed itself this past daylight savings (spring forward, in early-mid March) causing events to be displayed one hour off, if viewing the calendar on the day of the daylight savings change (Sunday). To be clear, this means the *current day while viewing was Sunday*, not that you were viewing Sunday on the calendar. As long as you were viewing the calendar during the Sunday of daylight savings, all events regardless of whether they were before, during, or after that day would be off by an hour.

Some updating of date/time calculations and rendering logic was needed to fix this bug.

Resolves:
https://elationhealth.atlassian.net/browse/RM-4985