Change href=# to href=javascript:void(0) to not break hash routing
Some folks choose to do client-side routing on a web client using the
window.location.hash property.  Manipulating after the # symbol does
not trigger a server request.  There are different ways to achieve
this now, but it is still a valid choice for folks.

This change edits all dummy `<a href="#"...` links to
`<a href="javascript:void(0)" ...` as the latter form achieves the
same results without messing up the client-side routing for folks using
the `window.location.hash` property.