DnD plugin: fully support custom event objects
While trying to use the drag and drop plugin with ImmutableJS
events I was getting a number of errors which were caused by
my custom accessors getting some close but not quite JS objects
rather than my glorious ImmutableJS Maps /s

Eventually I narrowed it down to the `update` function in
EventContainerWrapper - which in essence is creating a new
version of the event object in state with modified start & end
times.

Unfortunately it was simply doing this by treating the event as
a regular JS object with start & end attributes.  This is then
consumed by the TimeGridEvent component, which tries to use the
accessors again and then we're in trouble...

The fix involves creating customisable `updateEvent` function which
follows the same path as the `accessors` -- not 100% sure it's the
best, but it works for my purposes!

Also it could be that I'm just abusing the small amount of power
I was given with the accessors, but it felt a little odd to run into
an issue after having it work so well elsewhere!