Preserve time on horizontal resizing
When an event is resized horizontally (e.g. on the Month
View), keep the time of the event the same and only change
the date.

This also fixes the bug where you cannot resize an event to
a single day because the `dates.max` or `dates.min` would
force the event to occur at least 1 day after the
`start` or `end` respectively.